Build clock outline and shadow using canvas operations not image data
Get rid of our copyNonTransparentPixels() function, which copies pixels
manually from one ImageData object to another. It's slow. If you set the
outline size to something large (and by large I mean more than about
six), it takes a hundred milliseconds or so just to draw the outline.
This is because we draw the clock image in black outline_size^2 times to
make the outline, then we draw that outline image shadow_length times
to draw the shadow.

Instead, let's put the initial clock image into a canvas object, so we
can feed that to outlineCanvas.drawImage() to build the outline, then
feed that canvas to finishedCanvas.drawImage() to draw the shadow,
before we finally draw the actual clock canvas in white on to
finishedCanvas. This is much faster than copying all the pixels one
by one in a loop.