Option to show tenths of a second
Add a checkbox which if checked makes the clock show the time to one
decimal place, refreshing on every 100ms boundary.

Clock.formatValue() can now add an optional decimal point and
up to 3 decimal places, although there's currently no way in the
interface to make it display that many. You either get 0 or 1.