Some cosmetic changes to the menu, and redraw clock on window resize
The "count up" checkbox is now next to the minutes and seconds controls,
as it makes more sense to set up which direction you want the clock to
count at the same time as you set the initial time.

The Start/Stop button now goes green or red depending on which of those
two hats it happens to be wearing.

The Reset button is disabled if the clock is already reset (that is,
showing the initial time having not been started since it was last
reset). This gives the user one less irrelevant control to ponder over
when the page first loads.

The clock now gets automatically redrawn when you resize the window.

Some minor changes to the menu layout, such as the arrangement of the
colour pickers.