Add "show border" checkbox and make default background green
There is now a "show border" checkbox, with which the user can toggle
whether the octagonal border around the clock is shown. If it isn't
shown, we still reserve space in the image for it, so the numbers will
still be in the same place on the screen.

The default background colour is now #00b140, which I gather from at
least a dozen seconds of googling is known as "Chroma Key Green". This
replaces the previous background colour known to me as "Arbitrary Blue".