Export settings to URL
You can now get Octagon Simulator to give you a URL containing all the
currently-active settings in the query string. Anyone who visits the
URL will see the clock set up in the same way. You can also configure
the URL to automatically hide the menu or start the clock when the
page loads.

Also:

The windowSizeChanged() call is no longer in the <body> tag in the HTML
file - this caused errors if there was a resize event before the page
fully loaded and hence before windowSizeChanged was parsed. It's now
attached to window as an event listener.

refreshClock() can now cope with being called when clock == null, by
returning having done nothing. refreshClock can be called by a resize
event, which, as noted above, can occur before the page fully loads.