Checkbox to disable background colour entirely
This makes the clock marginally easier to use with OBS's Browser Source
feature, because you don't have to apply a chroma-key filter. OBS knows
what the background is and applies custom css to it
("background-color: rgba(0,0,0,0);") but that doesn't help if the page
sets the body's background colour in JavaScript which overrides it.

If you tick the new "disable background colour" box, the body tag's
background-color property is set to null regardless of what the
colour picker says, so any custom CSS takes effect.

Also, in initialisePost(), rearrange some of the code so that we
refreshConfiguration() and refreshClock() after applying any settings in
the URL, not before.