Introduce counter mode
Introduce a "counter mode", which uses the display as a counter rather
than a clock and allows the user to increment, decrement or reset the
counter.

Currently counter mode only supports a 2-digit octagon width, which is
widened as necessary for numbers of more than two digits.

There are keyboard shortcuts. The minus key decrements and the plus
(or equals) key increments.