Put the clock drawing code in its own class
Introduce the class CanvasClockDesign, which extends ClockDesign.
This encapsulates a clock design which paints into a canvas. A
CanvasClockDesign is instantiated with all the character image files
and measurements it needs.

Currently the only class which extends CanvasClockDesign is the
MicroCanvasClockDesign, which is the design that already exists.
There are no user-visible changes in behaviour. However, this change
will make it much easier to add alternative clock designs in the future.

font.js (generated by gendata.py) used to contain an array
characterUrls. Instead, it now contains a dictionary
characterDesignUrls. Each key is a design name (e.g. "micro") and it
maps to an array of URLs (like characterUrls), one per character.

gendata.py now takes each subdirectory of the png directory as a
separate clock design, and that directory must contain the expected .png
filenames.