Second clock design
Introduce a new clock design which uses HTML rather than painting pixels
on a canvas. The user can switch between this and the old one using the
"Design" drop-down box in the "Format & display" section. The default
is still the old one.

This one is designed to look a bit more modern. The font is Inter, which
supports the "font-variant-numeric: tabular-nums;" CSS property so we
can make the digits have a fixed width.

For the new design, the "border" checkbox enables and disables the
rectangular background of the new clock design. The "outline length" is
capped at 0.5 because we abuse the text-shadow property to do a text
outline and it doesn't work well for anything above that.