Separate drawClock() method from style and appearance methods
drawClock() currently takes the text to display on the clock and about a
million and one arguments specifying various style options such as text
colour, shadow length and so on.

Now that drawClock() is a method in a ClockDesign class rather than its
own standalone function, drawClock() should only be called to change the
reading on the clock. The ClockDesign object should remember its style
settings, which can be set by other methods such as setOutline(),
setPosition(), and so on.

This means that when refreshClock() is called, which could be up to ten
times a second, the HTML clock design can just change the contents of
the div rather than have to apply the style information all over again.
The style information only needs to be reapplied when something calls
the methods to change the clock design's style settings.

The "show border" option is moved from the "options" category to the
"appearance" category.