Introduce Nixie Tube clock design
Introduce images and code for the Nixie tube clock design.

The NixieCanvasClockDesign class extends CanvasClockDesign, so it's the
same sort of thing as the original MicroCanvasClockDesign, except that
the images don't have a transparent background. This means the text
outline and shadow options have no effect for this clock design.

gendata.py now copes with both png and jpg image formats. Rename the
png directory to images to reflect this.

Fix the comment describing borderToCharSpace, which was wrong.