Introduce presets window with suggested configurations
Introduce a presets window at the top of the menu, which suggests a
number of pre-canned configurations. Each one has a preview image and a
(very) short description.

Clicking a preset follows a link which reloads the page with the new
settings. At some point soon we probably want to change this so it
applies the new settings to the currently-running clock without
reloading the page.