Optimisations for drawing images on canvas
Use the canvas drawImage() method where possible rather than copying
pixels in an ImageData. This gives better performance.

Also: when we clear the canvas before every repaint, only clear the
region that's dirty if we can get away with that. In the event that
any style options have been changed, or the size of the canvas has been
changed since the last repaint, clear the whole thing.

When we build the clock image (which I'm now calling the "brush"
because we use it to draw the outline, shadow and text in various
colours), build it in an existing canvas rather than creating a new one
for every repaint only to throw it away.