Make preset buttons not reload the page
The preset buttons now load a pre-defined set of options for that preset
(defined in presets.js) rather than reloading the page.

The preset buttons are disabled while the clock is running, because it's
not necessarily obvious that changing to a different preset while the
clock is running might change the direction the clock runs in.

The Boring clock no longer uses bold text. I considered bold numbers to
be insufficiently boring.

The page now has a favicon.