Provide selection of typefaces for use with Boring design
Six font families are selected from the Google Fonts library for use
with the Boring design. They are Bebas Neue, Boogaloo, Inter, Nova Mono,
Press Start 2P, and Questrial. They're different styles of font but all
have or support fixed-width numbers (either with "font-variant-numeric:
tabular-nums;" or naturally).

The default is Inter, the one used by the Boring design before.

Tweak to take account of the fact that &numsp; is not always the same
width as a number.

Update About text.