Introduce Teletext clock design
TeletextCanvasClockDesign extends CanvasClockDesign and has a load of
little images to go with it. The images come from 12x10 pixel bitmaps
blown up to 60x50. The "border" images are just the same as the space,
with the result that the clock image begins and ends with a space.

The teletext font seems to be based on pixels which are 1.5 times as
high as they are wide, so a vertical stretch factor is introduced to
CanvasClockDesign. Normally this is 3/2 but if the new doubleHeight
property is set to true, it's 3.

Currently only the Teletext clock supports the double-height characters
option.

The Teletext clock also supports a background colour and transparency,
and CanvasClockDesign now knows how to do that (fillRect the canvas with
the appropriate colour).