Reorganise options, add teletext real-time clock preset, etc
Options are now reorganised into: "General", containing things that apply
to all designs and modes such as size, position, page background etc;
"Format", containing options on formatting the text that appears in the
clock (show tenths, etc); and "Design", containing design-specific
options.

If you use the Design drop-down to switch designs, the stuff in the
"Design" section automatically gets overwritten with that design's
default options, unless you switch this feature off. This is useful
if you loaded, say, the Teletext design using a preset, but then used
the drop-down to select Octagon - normally you're going to want
Octagon's default shadow and outline settings, not Teletext's.

Fix formatMilliseconds and formatValue in clock.js, which weren't
respecting the clock's direction correctly after the last commit. The
clock's direction is important because we want, say, "2.999" to be
shown as "3" in a countdown timer and "2" in a count-up timer.

Comment out the double-height Teletext presets, and instead let it
be a feature to be discovered in the menu. The presets bar was getting
too wide.

Put the keyboard shortcut cheat sheet at the bottom of the menu and have
it take up less space.