Improved mobile and touchscreen experience
The clock can now be auto-centred and maximised to no more than 70% of
either width or height, which is more suitable for display on a mobile
or tablet.

If we detect that the user is using a touchscreen, we start with the
main menu not displayed, the clock auto-centred and maximised, and the
touchscreen controls enabled. The touchscreen controls are buttons for
basic functionality along the bottom of the screen (if portrait) or down
the right edge of the screen (if landscape).

The touchscreen buttons available depend on the mode: in clock mode we
can start/stop and reset, and in counter mode we can increment,
decrement or reset. In all modes, the "hamburger" button is visible to
display the main menu.

The main menu has a phone icon at the top of the screen to manually
enable/disable the touchscreen controls should we guess wrong whether
the user wants them.

The main menu can now be closed with an X in the top right. In the event
that a user with a touchscreen and no keyboard closes it but doesn't
have the touchscreen controls to bring the menu back, there is now a new
way to bring up the menu: tap the screen anywhere in the left quarter of
the top half.

If the touchscreen buttons are enabled, there is also a "toggle
fullscreen" button. This is always in the top-right corner regardless of
whether the screen is portrait or landscape.