Fix HTML clock not sizing correctly on first display
When you switch from the default clock to the Boring HTML clock for the
first time while it's centred and maximised, it puts it in slightly the
wrong place. If you click the Boring preset button again, it puts it in
the right place.

Fixed by initialising the HTML clock div to something non-empty.
Apparently clientWidth/clientHeight don't get calculated correctly if
there was no text in the div before.